requirements.txt
requests
pillow
streamlit
